+++ qiskit/__init__.py
@@ -28,6 +28,8 @@ from .tools.compiler import (compile, execute)
 # Please note these are global instances, not modules.
 from qiskit.backends.ibmq import IBMQ
 from qiskit.backends.aer import Aer  # pylint: disable=invalid-name
+from qiskit.backends.aer_py import AerPy # pylint: disable=invalid-name
+
 
 # The qiskit.extensions.x imports needs to be placed here due to the
 # mechanism for adding gates dynamically.
