+++ qiskit/backends/aer/__init__.py
@@ -9,11 +9,8 @@
 
 from .aerprovider import AerProvider
 from .aerjob import AerJob
-from .qasm_simulator import CliffordSimulator, QasmSimulator
-from .qasm_simulator_py import QasmSimulatorPy
-from .statevector_simulator import StatevectorSimulator
-from .statevector_simulator_py import StatevectorSimulatorPy
-from .unitary_simulator_py import UnitarySimulatorPy
+from .qasmsimulator import CliffordSimulator, QasmSimulator
+from .statevectorsimulator import StatevectorSimulator
 
 # Global instance to be used as the entry point for convenience.
 Aer = AerProvider()  # pylint: disable=invalid-name
