+++ qiskit/backends/aer/statevectorsimulator.py
@@ -7,46 +7,42 @@
 
 # pylint: disable=invalid-name
 
-"""Contains a (slow) python statevector simulator.
-
-It simulates the statevector through a quantum circuit. It is exponential in
-the number of qubits.
-
-We advise using the c++ simulator or online simulator for larger size systems.
-
-The input is a qobj dictionary and the output is a Result object.
-
-The input qobj to this simulator has no shots, no measures, no reset, no noise.
 """
+Interface to C++ quantum circuit simulator with realistic noise.
+"""
+
 import logging
 import uuid
 from math import log2
+from numpy import array
 from qiskit._util import local_hardware_info
-from qiskit.backends.aer.aerjob import AerJob
-from qiskit.backends.aer._simulatorerror import SimulatorError
 from qiskit.backends.models import BackendConfiguration
 from qiskit.qobj import QobjInstruction
-from .qasm_simulator_py import QasmSimulatorPy
+from .qasmsimulator import QasmSimulator
+from .aererror import AerError
+from .aerjob import AerJob
 
 logger = logging.getLogger(__name__)
 
 
-class StatevectorSimulatorPy(QasmSimulatorPy):
-    """Python statevector simulator."""
+class StatevectorSimulator(QasmSimulator):
+    """C++ statevector simulator"""
 
     DEFAULT_CONFIGURATION = {
-        'backend_name': 'statevector_simulator_py',
+        'backend_name': 'statevector_simulator',
         'backend_version': '1.0.0',
         'n_qubits': int(log2(local_hardware_info()['memory'] * (1024**3)/16)),
-        'url': 'https://github.com/Qiskit/qiskit-terra',
+        'url': 'https://github.com/Qiskit/qiskit-terra/src/qasm-simulator-cpp',
         'simulator': True,
         'local': True,
         'conditional': False,
         'open_pulse': False,
         'memory': False,
         'max_shots': 65536,
-        'description': 'A Python statevector simulator for qobj files',
-        'basis_gates': ['u1', 'u2', 'u3', 'cx', 'id', 'snapshot'],
+        'description': 'A single-shot C++ statevector simulator for the |0> state evolution',
+        'basis_gates': ['u1', 'u2', 'u3', 'cx', 'cz', 'id', 'x', 'y', 'z', 'h',
+                        's', 'sdg', 't', 'tdg', 'rzz', 'load', 'save',
+                        'snapshot'],
         'gates': [
             {
                 'name': 'u1',
@@ -69,9 +65,69 @@ class StatevectorSimulatorPy(QasmSimulatorPy):
                 'qasm_def': 'gate cx c,t { CX c,t; }'
             },
             {
+                'name': 'cz',
+                'parameters': ['a', 'b'],
+                'qasm_def': 'gate cz a,b { h b; cx a,b; h b; }'
+            },
+            {
                 'name': 'id',
                 'parameters': ['a'],
                 'qasm_def': 'gate id a { U(0,0,0) a; }'
+            },
+            {
+                'name': 'x',
+                'parameters': ['a'],
+                'qasm_def': 'gate x a { u3(pi,0,pi) a; }'
+            },
+            {
+                'name': 'y',
+                'parameters': ['a'],
+                'qasm_def': 'gate y a { u3(pi,pi/2,pi/2) a; }'
+            },
+            {
+                'name': 'z',
+                'parameters': ['z'],
+                'qasm_def': 'gate z a { u1(pi) a; }'
+            },
+            {
+                'name': 'h',
+                'parameters': ['a'],
+                'qasm_def': 'gate h a { u2(0,pi) a; }'
+            },
+            {
+                'name': 's',
+                'parameters': ['a'],
+                'qasm_def': 'gate s a { u1(pi/2) a; }'
+            },
+            {
+                'name': 'sdg',
+                'parameters': ['a'],
+                'qasm_def': 'gate sdg a { u1(-pi/2) a; }'
+            },
+            {
+                'name': 't',
+                'parameters': ['a'],
+                'qasm_def': 'gate t a { u1(pi/4) a; }'
+            },
+            {
+                'name': 'tdg',
+                'parameters': ['a'],
+                'qasm_def': 'gate tdg a { u1(-pi/4) a; }'
+            },
+            {
+                'name': 'rzz',
+                'parameters': ['theta', 'a', 'b'],
+                'qasm_def': 'gate rzz(theta) a,b { cx a,b; u1(theta) b; cx a,b; }'
+            },
+            {
+                'name': 'load',
+                'parameters': ['slot'],
+                'qasm_def': 'gate load(slot) q { TODO }'
+            },
+            {
+                'name': 'save',
+                'parameters': ['slot'],
+                'qasm_def': 'gate save(slot) q { TODO }'
             },
             {
                 'name': 'snapshot',
@@ -87,14 +143,7 @@ class StatevectorSimulatorPy(QasmSimulatorPy):
                          provider=provider)
 
     def run(self, qobj):
-        """Run qobj asynchronously.
-
-        Args:
-            qobj (dict): job description
-
-        Returns:
-            AerJob: derived from BaseJob
-        """
+        """Run a qobj on the backend."""
         job_id = str(uuid.uuid4())
         aer_job = AerJob(self, job_id, self._run_job, qobj)
         aer_job.submit()
@@ -117,11 +166,11 @@ class StatevectorSimulatorPy(QasmSimulatorPy):
         # Extract final state snapshot and move to 'statevector' data field
         for experiment_result in result.results:
             snapshots = experiment_result.data.snapshots.to_dict()
-            if str(final_state_key) in snapshots:
+            if str(final_state_key) in snapshots['statevector']:
                 final_state_key = str(final_state_key)
             # Pop off final snapshot added above
-            final_state = snapshots.pop(final_state_key, None)
-            final_state = final_state['statevector'][0]
+            final_state = snapshots['statevector'].pop(final_state_key)[0]
+            final_state = array([v[0] + 1j * v[1] for v in final_state], dtype=complex)
             # Add final state to results data
             experiment_result.data.statevector = final_state
             # Remove snapshot dict if empty
@@ -147,6 +196,6 @@ class StatevectorSimulatorPy(QasmSimulatorPy):
                 experiment.config.shots = 1
             for op in experiment.instructions:
                 if op.name in ['measure', 'reset']:
-                    raise SimulatorError(
+                    raise AerError(
                         "In circuit {}: statevector simulator does not support "
                         "measure or reset.".format(experiment.header.name))
