+++ qiskit/backends/aer_py/__init__.py
@@ -7,13 +7,11 @@
 
 """Aer Backends."""
 
-from .aerprovider import AerProvider
-from .aerjob import AerJob
-from .qasm_simulator import CliffordSimulator, QasmSimulator
-from .qasm_simulator_py import QasmSimulatorPy
-from .statevector_simulator import StatevectorSimulator
-from .statevector_simulator_py import StatevectorSimulatorPy
-from .unitary_simulator_py import UnitarySimulatorPy
+from .aerpyprovider import AerPyProvider
+from .aerpyjob import AerPyJob
+from .qasmsimulator import QasmSimulatorPy
+from .statevectorsimulator import StatevectorSimulatorPy
+from .unitarysimulator import UnitarySimulatorPy
 
 # Global instance to be used as the entry point for convenience.
-Aer = AerProvider()  # pylint: disable=invalid-name
+AerPy = AerPyProvider()  # pylint: disable=invalid-name
