+++ qiskit/backends/aer_py/aerpyprovider.py
@@ -6,43 +6,38 @@
 # the LICENSE.txt file in the root directory of this source tree.
 
 
-"""Provider for aer backends."""
+"""Provider for aer python backends."""
 
 from collections import OrderedDict
 import logging
 
-from qiskit._qiskiterror import QiskitError
+from qiskit.qiskiterror import QiskitError
 from qiskit.backends import BaseProvider
 from qiskit.backends.exceptions import QiskitBackendNotFoundError
 from qiskit.backends.providerutils import resolve_backend_name, filter_backends
 
-from .qasm_simulator import CliffordSimulator, QasmSimulator
-from .qasm_simulator_py import QasmSimulatorPy
-from .statevector_simulator import StatevectorSimulator
-from .statevector_simulator_py import StatevectorSimulatorPy
-from .unitary_simulator_py import UnitarySimulatorPy
+from .qasmsimulator import QasmSimulatorPy
+from .statevectorsimulator import StatevectorSimulatorPy
+from .unitarysimulator import UnitarySimulatorPy
 
 
 logger = logging.getLogger(__name__)
 
-AER_STANDARD_BACKENDS = [
-    QasmSimulator,
+AERPY_STANDARD_BACKENDS = [
     QasmSimulatorPy,
-    StatevectorSimulator,
     StatevectorSimulatorPy,
     UnitarySimulatorPy,
-    CliffordSimulator,
 ]
 
 
-class AerProvider(BaseProvider):
-    """Provider for aer backends."""
+class AerPyProvider(BaseProvider):
+    """Provider for aer pyhton backends."""
 
     def __init__(self, *args, **kwargs):
         super().__init__(args, kwargs)
 
-        # Populate the list of aer backends.
-        self._backends = self._verify_aer_backends()
+        # Populate the list of aer python backends.
+        self._backends = self._verify_aer_py_backends()
 
     def get_backend(self, name=None, **kwargs):
         backends = self._backends.values()
@@ -55,7 +50,7 @@ class AerProvider(BaseProvider):
                     name, backends,
                     self._deprecated_backend_names(),
                     {},
-                    self._alternative_py_backend_names()
+                    {}
                 )
                 name = resolved_name
             except LookupError:
@@ -75,8 +70,7 @@ class AerProvider(BaseProvider):
                 resolved_name = resolve_backend_name(
                     name, backends,
                     self._deprecated_backend_names(),
-                    {},
-                    self._alternative_py_backend_names()
+                    {}, {}
                 )
                 backends = [backend for backend in backends if
                             backend.name() == resolved_name]
@@ -89,37 +83,31 @@ class AerProvider(BaseProvider):
     def _deprecated_backend_names():
         """Returns deprecated backend names."""
         return {
-            'local_qasm_simulator_cpp': 'qasm_simulator',
-            'local_qasm_simulator_py': 'qasm_simulator_py',
-            'local_statevector_simulator_cpp': 'statevector_simulator',
-            'local_statevector_simulator_py': 'statevector_simulator_py',
-            'local_unitary_simulator_py': 'unitary_simulator_py',
-            'local_qiskit_simulator': 'qasm_simulator',
-            'local_qasm_simulator': 'qasm_simulator',
-            'local_statevector_simulator': 'statevector_simulator',
-            'local_unitary_simulator': 'unitary_simulator_py',
-            'unitary_simulator': 'unitary_simulator_py'
+            'local_qasm_simulator_py': 'qasm_simulator',
+            'local_statevector_simulator_py': 'statevector_simulator',
+            'local_unitary_simulator_py': 'unitary_simulator',
+            'local_unitary_simulator': 'unitary_simulator',
             }
 
-    def _verify_aer_backends(self):
+    def _verify_aer_py_backends(self):
         """
-        Return the aer backends in `AER_STANDARD_BACKENDS` that are
+        Return the aer python backends in `AERPY_STANDARD_BACKENDS` that are
         effectively available (as some of them might depend on the presence
         of an optional dependency or on the existence of a binary).
 
         Returns:
-            dict[str:BaseBackend]: a dict of aer backend instances for
+            dict[str:BaseBackend]: a dict of aer  python backend instances for
                 the backends that could be instantiated, keyed by backend name.
         """
         ret = OrderedDict()
-        for backend_cls in AER_STANDARD_BACKENDS:
+        for backend_cls in AERPY_STANDARD_BACKENDS:
             try:
                 backend_instance = self._get_backend_instance(backend_cls)
                 backend_name = backend_instance.name()
                 ret[backend_name] = backend_instance
             except QiskitError as err:
                 # Ignore backends that could not be initialized.
-                logger.info('aer backend %s is not available: %s',
+                logger.info('aer python backend %s is not available: %s',
                             backend_cls, str(err))
         return ret
 
@@ -144,12 +132,4 @@ class AerProvider(BaseProvider):
         return backend_instance
 
     def __str__(self):
-        return 'Aer'
-
-    @staticmethod
-    def _alternative_py_backend_names():
-        """Return Python alternatives to C++ simulators."""
-        return {
-            'qasm_simulator': 'qasm_simulator_py',
-            'statevector_simulator': 'statevector_simulator_py'
-        }
+        return 'AerPy'
